EduInsights/backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
pandas
scikit-learn
joblib
numpy
python-dotenv
gunicorn
whitenoise
django-cors-headers
